excel增加字段及行宽、列高修改

+++ kanban-api-commons/src/main/java/com/honortone/commons/entity/Product.java
@@ -36,13 +36,13 @@ public class Product implements Serializable {
     private String uid;
 
     @ApiModelProperty("数量")
-    private Integer quantity;
+    private Double quantity;
 
     @ApiModelProperty("库位号")
     private String storageLoc;
 
     @ApiModelProperty("状态")
-    private Integer state;
+    private Short state;
 
     @ApiModelProperty("工单")
     private String wo;
